pyproject.toml
[tool.poetry]
name = "glossary"
version = "0.2"
description = "Python package to define technical terms"
authors = ["ian-flores <iflores.siaca@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3.1"
textdistance = "^4.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"